Added Village and Smithy action cards
Super terrible actions are implemented, this should be fine for testing.
Also fixed a bug where it crashed when you tried to buy nothing by
inputting "none".
Also added the notion of "in play" to Cards, this should probably be
useful for future kinds of action cards (i.e. orange cards).

Notes:
- If you play "all" cards and you have actions that aren't before money
  cards, the money cards get played and your actions are skipped (and
  you no longer have your starting +1 action) this is WAI although a
  warning against playing "all"
- To finish a turn you must type "done"
- You can still play victory point cards, this still should be fixed